Add core app, choices.py, utils.py
Исправил замечания, чекайте
