+++ core/choices.py
@@ -1,0 +1,9 @@
+STATUS_CHOICES = [
+        ('active', 'Активные'),
+        ('completed', 'Выполненные'),
+        ('two_weeks', 'Две недели до плановой даты выполнения'),
+        ('overdue', 'Просроченные'),
+        ('canceled', 'Отмененные'),
+        ('closed', 'Закрытые'),
+        ('draft', 'Черновик')
+    ]
