+++ core/utils.py
@@ -1,0 +1,6 @@
+from datetime import timedelta
+from django.utils import timezone
+
+
+def default_end_date_plan():
+    return timezone.now() + timedelta(days=30)
